requirements.txt
Flask==3.1.0
pymongo[srv]==4.11.3
python-dotenv==1.0.1
mongomock==4.3.0
gunicorn==23.0.0
